@@ -3,28 +3,30 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@material-ui/core": "^4.11.3",
-		"@material-ui/icons": "^4.11.2",
-		"@testing-library/jest-dom": "^5.11.9",
-		"@testing-library/react": "^11.2.5",
-		"@testing-library/user-event": "^12.6.3",
-		"@types/react-router-dom": "^5.1.7",
-		"@types/socket.io-client": "^1.4.35",
-		"@types/styled-components": "^5.1.7",
-		"axios": "^0.21.1",
-		"date-fns": "^2.17.0",
-		"react": "^17.0.1",
-		"react-debounce-input": "^3.2.3",
-		"react-dom": "^17.0.1",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "4.0.1",
-		"react-scrollbars-custom": "^4.0.25",
-		"rxjs": "^6.6.3",
-		"socket.io-client": "3.1.1",
-		"styled-components": "^5.2.1",
+		"@emotion/react": "^11.11.4",
+		"@emotion/styled": "^11.11.5",
+		"@mui/icons-material": "^5.13.7",
+		"@mui/material": "^5.13.7",
+		"@testing-library/jest-dom": "^5.16.5",
+		"@testing-library/react": "^13.4.0",
+		"@testing-library/user-event": "^14.4.3",
+		"@types/styled-components": "^5.1.25",
+		"axios": "^1.3.4",
+		"date-fns": "^2.30.0",
+		"react": "^18.2.0",
+		"react-custom-scrollbars-2": "^4.5.0",
+		"react-debounce-input": "^3.3.0",
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.23.1",
+		"react-scripts": "^5.0.1",
+		"react-scrollbars-custom": "^4.0.34",
+		"react-stomp-hooks": "^2.2.1",
+		"rxjs": "^7.8.1",
+		"socket.io-client": "^4.7.5",
+		"styled-components": "^5.3.10",
 		"sweetalert2": "^11.11.0",
-		"typescript": "^4.1.3",
-		"web-vitals": "^0.2.4"
+		"typescript": "^4.9.5",
+		"web-vitals": "^3.1.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -38,16 +40,9 @@
 			"react-app/jest"
 		]
 	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	}
+	"browserslist": [
+		">0.2%",
+		"not dead",
+		"not op_mini all"
+	]
 }
